# documentation/swagger.yaml
openapi: 3.0.0
info:
  title: MamaCheck API - Complete Reference
  description: |
    # MamaCheck - SMS-First Antenatal Care System

    Comprehensive REST API for pregnancy management, ANC reminders, and danger sign triage in Nigeria.

    ## Key Features
    - **Pregnancy Management**: Registration with OTP verification, timeline tracking
    - **ANC Reminders**: Daily scheduling at 07:00 WAT, personalized messages
    - **Danger Sign Triage**: Weekly check-ins with RED/YELLOW/GREEN outcomes
    - **CHEW Dashboard**: Real-time KPIs, women registry, red flag management
    - **Multi-language**: English, Pidgin, Yoruba, Hausa, Igbo
    - **SMS Integration**: Termii SMS gateway with DND compliance
    - **Visit Tracking**: Attendance marking with 10-minute undo window
    - **Reference Data**: LGA and PHC management with geolocation

    ## Security
    - JWT authentication for all protected endpoints
    - Role-based access (PATIENT, CHEW, SUPERVISOR, ADMIN)
    - OTP verification for phone number validation
    - Rate limiting on SMS and registration endpoints
    - BCRYPT password hashing in production

    ## SMS Triage Guide
    Women reply with symptom numbers:
    - **0** = No symptoms (GREEN)
    - **1** = Heavy bleeding (RED)
    - **2** = Severe headache (RED)
    - **3** = Swollen face/hands (RED)
    - **4** = Blurry vision (YELLOW)
    - **5** = Fever (YELLOW)
    - **6** = Reduced baby movement (YELLOW)
...
  version: "1.0.0"
...
servers:
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticate with phone and password. Returns JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - password
              properties:
                phone:
                  type: string
                  example: "08012345678"
                  description: Nigerian phone number
                password:
                  type: string
                  format: password
                  example: "securePassword123"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authenticated requests
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      phone:
                        type: string
                      role:
                        type: string
                        enum: [patient, chew, supervisor, admin]
        "401":
          description: Invalid phone or password
        "429":
          description: Too many login attempts - rate limited
  /auth/request-otp:
    post:
      tags:
        - Authentication
      summary: Request OTP
      description: Send OTP via SMS to phone number. Requires active pregnancy for women.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  example: "08012345678"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent successfully"
        "500":
          description: Failed to send OTP (Termii error)
  /auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP
      description: Verify OTP and receive authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - otp
              properties:
                phone:
                  type: string
                  example: "08012345678"
                otp:
                  type: string
                  example: "123456"
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: OTP verified, token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        "401":
          description: Invalid or expired OTP
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get Current User
      description: Retrieve authenticated user details
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [patient, chew, supervisor, admin]
        preferredLanguage:
          type: string
          enum: [en, pidgin, yo, ha, ig]
        trustedContact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
        consent:
          type: object
          properties:
            sms:
              type: boolean
        optOut:
          type: object
          properties:
            isOptedOut:
              type: boolean
...